+++ modules/gui/gnome.nix
@@ -1,0 +1,7 @@
+{ config, pkgs, lib, ... }:
+
+{
+  services.xserver = {
+    desktopManager.gnome.enable = true;
+  };
+}