+++ modules/keymap.nix
@@ -1,0 +1,12 @@
+{ config, pkgs, lib, ... }:
+
+{
+  # Configure keymap in X11
+  services.xserver.xkb = {
+    layout = "fr";
+    variant = "";
+  };
+
+  # Configure console keymap
+  console.keyMap = "fr";
+}